database: Move the static methods into a separate file.
